Tubes: Refactoring, Map and Types
