+++ libopenage/CMakeLists.txt
@@ -46,7 +46,7 @@ add_subdirectory("rng")
 add_subdirectory("shader")
 add_subdirectory("terrain")
 add_subdirectory("testing")
-add_subdirectory("tubes")
+add_subdirectory("tube")
 add_subdirectory("unit")
 add_subdirectory("util")
