+++ libopenage/tube/map_filter_iterator.h
@@ -10,6 +10,18 @@
 namespace openage {
 namespace tube {
 
+template <typename val_t>
+class TubeIterator {
+public:
+	virtual val_t &value() = 0;
+};
+
+template<typename _T>
+bool valid(const _T &, const tube_time_t &at);
+
+template <class _T>
+using _valid_function_t = bool (*)(const _T&, const tube_time_t &);
+
 template <class key_t,
           class val_t,
           class container_t,
@@ -26,6 +38,7 @@ protected:
 
 	tube_time_t from;
 	tube_time_t to;
+	tube_time_t now;
 
 public:
 	TubeMapFilterIterator(const iterator_t &base,
@@ -36,12 +49,14 @@ public:
 		container_end{container_end},
 		from{from},
 		to{to} {}
+		now{from} {}
 
 	TubeMapFilterIterator(const TubeMapFilterIterator &rhs) :
 		base{rhs.base},
 		container_end{rhs.container_end},
 		from{rhs.from},
-		to{rhs.to} {}
+		to{rhs.to},
+		now{rhs.now} {}
 
 	TubeMapFilterIterator &operator =(const TubeMapFilterIterator &rhs) {
 		this->base = rhs.base;
@@ -95,5 +110,12 @@ public:
 		return this->base->first;
 	}
 };
+
+template<typename _T>
+bool valid(const _T &t,
+           const tube_time_t& time) {
+	return existent_from(t) <= time && existent_until(t) > time;
+}
+
 
 }} // openage::tube
