--- libopenage/tubes/README.md
@@ -1,5 +1,0 @@
-# tubepong
-
-An experiment how one can work with tubes and predictionmagic in a stateless physical game engine.
-
-This is a Pong-like game build on top of tubes.
