--- libopenage/tubes/demo/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_sources (libopenage
-	main.cpp
-	physics.cpp
-	gui.cpp
-	aicontroller.cpp
-)
