--- libopenage/tubes/event.cpp
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "event.h"
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
