--- libopenage/tubes/filter_iterator.cpp
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "filter_iterator.h"
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
