--- libopenage/tubes/object.h
@@ -1,14 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#pragma once
-
-#include "tube.h"
-
-namespace openage {
-namespace tube {
-
-class TubeObject {
-public:
-};
-
-}} // openage::tube
