--- libopenage/tubes/objectlist.h
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#pragma once
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
