--- libopenage/tubes/simple_discrete.cpp
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "simple_discrete.h"
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
