--- libopenage/tubes/simple_type.cpp
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "simple_type.h"
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
