--- libopenage/tubes/test/test_serialization.cpp
@@ -1,13 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "../../testing/testing.h"
-
-namespace openage {
-namespace tube {
-namespace tests {
-
-void serialization() {
-
-}
-
-}}} // openage::tube::tests
