--- libopenage/tubes/test/test_tube_types.cpp
@@ -1,12 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "../../testing/testing.h"
-
-namespace openage {
-namespace tube {
-namespace tests {
-
-void tube_types() {
-}
-
-}}} // openage::tube::tests
