--- libopenage/tubes/tube.h
@@ -1,10 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#pragma once
-
-namespace openage {
-namespace tube {
-
-typedef float tube_time_t;
-
-}} // openage::tube
