--- libopenage/tubes/unordered_map.cpp
@@ -1,8 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#include "unordered_map.h"
-
-namespace openage {
-namespace tube {
-
-}} // openage::tube
